HIPP-1318: Upgrade to Scala3 (#62)
* HIPP-1318: Upgrade to Scala3

* HIPP-1318: Address PR comments

+++ app/uk/gov/hmrc/integrationcatalogue/config/AppConfig.scala
@@ -28,7 +28,7 @@ class AppConfig @Inject() (config: Configuration) {
 
   val defaultShortDescLength: Int   = 180
   val oldIndexesToDrop: Seq[String] = config.getOptional[Seq[String]]("mongodb.oldIndexesToDrop").getOrElse(Seq.empty)
-  val apiTeamsTtlDays: Int = config.getOptional[Int]("mongodb.apiTeamsTtlDays").getOrElse(180)
+  val apiTeamsTtlDays: Long = config.getOptional[Long]("mongodb.apiTeamsTtlDays").getOrElse(180)
   val shortDescLength: Int = config.getOptional[Int]("publish.shortDesc.maxLength").getOrElse(defaultShortDescLength)
 
   val authorizationKey: String = getString("authorizationKey")
